requirements.txt
streamlit
numpy
pandas
scipy
scikit-learn
underthesea
joblib
Pillow
